Configuring lombok

+++ src/main/java/com/juan/estevez/app/commons/GenericServiceImp.java
@@ -55,7 +55,10 @@ public abstract class GenericServiceImp<T, ID extends Serializable> implements G
 	 */
 	@Override
 	public void delete(ID id) {
-		getRepository().deleteById(id);
+		Optional<T> obj = getRepository().findById(id);
+		if (obj.isPresent()) {
+			getRepository().deleteById(id);
+		}
 	}
 
 	/**
